Evaluate an expression within context

Uses eval_tidy() to evaluate an expression within the context of data, however it gives much more flexibility to what can be supplied as an expression. See examples for a non-exhaustive list of uses. Beware, that the flexibility of eval_expr() may cause unexpected behavior in some circumstances.

eval_expr(expr, data, dot_expression = F, verbose = F)

Arguments

expr
expression to evaluate. Can be a call or a function, the name of a function, a quosure, a quote.

data
A data frame or named list or vector to provide context within which to evaluate expr

dot_expression
logical indicating whether to force a . expression evaluation. If . is at the top layer (e.g. length(.)), then this will be detected automatically

verbose
logical indicating whether to print a detailed description of what eval_expr() is doing.

Examples

#Basic evaluation in a data context
tbl <- tibble::tibble(x=1:100)
eval_expr(mean(x),tbl)

#> [1] 50.5

#Can also use functions defined by pipelines (fseq)
eval_expr(. %>% dplyr::pull(x) %>% mean,tbl)

#> [1] 50.5

#Different to eval_tidy() is that the data doesn't need names
eval_expr(mean,1:100)

#> [1] 50.5

x <- 1:50
eval_expr(mean,x)

#> [1] 25.5

#functions using dot notation work fine too
eval_expr(mean(.),1:100)

#> [1] 50.5

eval_expr(mean(.$x),tbl)

#> [1] 50.5

eval_expr(quantile(.,c(0.05,0.5,0.95)),1:100)

#>    5%   50%   95% 
#>  5.95 50.50 95.05 

#However, only simple cases are detected automatically
if (FALSE) {
eval_expr(mean(.)/length(.),1:100)
}

#More complicated expressions need to be explicit:
eval_expr(mean(.)/length(.),1:100,dot_expression=TRUE)

#> [1] 0.505

#It even pulls the function from the data
lst <- list(x = 1:100,fun = mean)
eval_expr(fun(x),lst)

#> [1] 50.5

#Be wary as functions defined within the data will mask those in
# the environment
lst <- list(x=1:100,mean = sd)
eval_expr(mean(x),lst)

#> [1] 29.01149

#It can even unquote arguments
command <- quote(mean(x))
eval_expr(command,tbl)

#> [1] 50.5

#Although this doesn't work explicitly
if (FALSE) {
eval_expr(quote(mean(x)),tbl)
}

#Quosures aren't a problem
my_quo <- rlang::quo(mean(x))
eval_expr(my_quo,tbl)

#> [1] 50.5

#And even within functions
f <- function(expr,data)
{
  quasi_expr <- rlang::enquo(expr)
  eval_expr(!!quasi_expr,data)
}

f(mean(x),tbl)

#> [1] 50.5

#And allows an environment to be passed as `data`

my_env <- new.env()
my_env$x <- tbl$x
eval_expr(mean(x),my_env)

#> [1] 50.5